chore: use async startup on node
