+++ apps/node-host/webpack.config.js
@@ -20,12 +20,12 @@ module.exports = composePlugins(withNx(), async (config) => {
       runtimePlugins: [
         require.resolve('@module-federation/node/runtimePlugin'),
       ],
+      experiments: {
+        asyncStartup: true,
+      },
       remotes: {
         node_local_remote:
           'commonjs ../../node-local-remote/dist/remoteEntry.js',
-        // node_local_remote: '__webpack_require__.federation.instance.moduleCache.get("node_local_remote")',
-        // node_remote:
-        //   '__webpack_require__.federation.instance.moduleCache.get("node_remote")@http://localhost:3002/remoteEntry.js',
         node_remote: 'node_remote@http://localhost:3022/remoteEntry.js',
       },
     }),
